Remove thredds-test dependence from TestResolve. * TestResolve.testResolver() used to depend on TestDir.threddsTestServer; it now depends on the embedded TDS. The feature collection used in the test has been changed to one on the embedded TDS. * In order to make that work, the test had to be moved from :cdm to :it. I created a new class with the same name in :it to house it. * Converted log output from STDOUT to slf4j.
